+++ src/presentation/repl-session-init.lisp
@@ -24,6 +24,9 @@
         *functions* (make-hash-table :test #'equal)
         *function-sources* (make-hash-table :test #'equal)
         *proc-registry* (make-hash-table :test #'eql))
+  (setf *vi-mode-enabled*
+        (let ((flag (uiop:getenv "NSHELL_VI_MODE")))
+          (and flag (not (member flag '("" "0" "false" "no") :test #'string-equal)))))
   (install-expansion-filesystem)
   (configure-completion-filesystem)
   (load-history-into-repl)